requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.118.0
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.3
opencv-python==4.8.1.78
packaging==25.0
pandas==2.3.2
pillow==11.3.0
polars==1.33.1
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
scipy==1.11.2
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.25
starlette==0.48.0
sympy==1.14.0
thop==0.1.1.post2209072238
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.203
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.37.0
